requirements.txt
requests>=2.31.0
Pillow>=10.0.0
tk>=0.1.0
python-dotenv>=1.0.0
pyinstaller>=6.3.0 